AIRFLOW-2045: Fixed bug in S3FileTransformOperator
`S3FileTransformOperator` was calling `get_contents_to_file`
which doesn't exist in boto3 >= 1.3. Trying to use earlier
versions of boto3 results in an error being thrown from
`S3Hook` trying to use a different method that doesn't exist
in boto3 < 1.3. `download_file` seems like the appropriate
method, and it requires the file path as a string, rather
than the file object.

The operator was also calling `connection.close()`, but
`connection` isn't an attribute in `S3Hook` and switching
it to `get_conn().close()` shows there's no `close()` method
in that S3 object.